requirements.txt
python-dotenv==0.21.0
google-genai==1.3.0
google-generativeai==0.4.1